Bug 636750 (3/4) - Use a double value instead of a float in nsAttrValue::MiscContainer. r=bz
UltraBlame original commit: f3c80cecaf9a187972ae85130bb1d945b767602a

+++ content/base/src/nsAttrValue.h
@@ -165,7 +165,7 @@ public:
 #ifdef MOZ_SVG
   inline nsISVGValue* GetSVGValue() const;
 #endif
-  inline float GetFloatValue() const;
+  inline double GetFloatValue() const;
   PRBool GetIntMarginValue(nsIntMargin& aMargin) const;
 
   
@@ -346,7 +346,7 @@ private:
 #ifdef MOZ_SVG
       nsISVGValue* mSVGValue;
 #endif
-      float mFloatValue;
+      double mFloatValue;
       nsIntMargin* mIntMargin;
     };
   };
@@ -458,7 +458,7 @@ nsAttrValue::GetSVGValue() const
 }
 #endif
 
-inline float
+inline double
 nsAttrValue::GetFloatValue() const
 {
   NS_PRECONDITION(Type() == eFloatValue, "wrong type");
